Moved PostsList toolbar from activity to fragment.
