+++ WordPress/src/main/java/org/wordpress/android/ui/posts/PostsListActivity.java
@@ -1,14 +1,12 @@
 package org.wordpress.android.ui.posts;
 
 import android.app.AlertDialog;
-import android.app.FragmentManager;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
+import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.text.TextUtils;
 import android.view.MenuItem;
 
@@ -17,6 +15,10 @@ import org.wordpress.android.WordPress;
 import org.wordpress.android.ui.ActivityId;
 import org.wordpress.android.ui.ActivityLauncher;
 import org.wordpress.android.util.ToastUtils;
+
+/*
+ * Serves as the host for PostsListFragment when showing uploaded posts.
+ */
 
 public class PostsListActivity extends AppCompatActivity {
     public static final String EXTRA_VIEW_PAGES = "viewPages";
@@ -31,22 +33,11 @@ public class PostsListActivity extends AppCompatActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
         setContentView(R.layout.post_list_activity);
 
         mIsPage = getIntent().getBooleanExtra(EXTRA_VIEW_PAGES, false);
 
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        ActionBar actionBar = getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.setTitle(getString(mIsPage ? R.string.pages : R.string.posts));
-            actionBar.setDisplayShowTitleEnabled(true);
-            actionBar.setDisplayHomeAsUpEnabled(true);
-        }
-
-        FragmentManager fm = getFragmentManager();
+        FragmentManager fm = getSupportFragmentManager();
         mPostList = (PostsListFragment) fm.findFragmentById(R.id.postList);
 
         showErrorDialogIfNeeded(getIntent().getExtras());
@@ -84,9 +75,9 @@ public class PostsListActivity extends AppCompatActivity {
 
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle(getResources().getText(R.string.error))
-               .setMessage(errorMessage)
-               .setPositiveButton(R.string.ok, null)
-               .setCancelable(true);
+                .setMessage(errorMessage)
+                .setPositiveButton(R.string.ok, null)
+                .setCancelable(true);
 
         // enable browsing error link if one exists
         if (!TextUtils.isEmpty(errorInfoTitle) && !TextUtils.isEmpty(errorInfoLink)) {
