+++ WordPress/src/main/res/layout/post_list_activity.xml
@@ -1,17 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<fragment android:id="@+id/postList"
+    android:name="org.wordpress.android.ui.posts.PostsListFragment"
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
+    android:layout_height="match_parent" />
 
-    <include
-        android:id="@+id/toolbar"
-        layout="@layout/toolbar" />
 
-    <fragment
-        android:id="@+id/postList"
-        android:name="org.wordpress.android.ui.posts.PostsListFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-</LinearLayout>
