+++ WordPress/src/main/res/layout/post_list_fragment.xml
@@ -1,85 +1,95 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/root_view"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
-    <LinearLayout
-        android:id="@+id/empty_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:animateLayoutChanges="true"
-        android:background="@color/background_grey"
-        android:gravity="center"
-        android:orientation="vertical"
-        android:visibility="gone"
-        tools:visibility="visible">
+    <include
+        android:id="@+id/toolbar"
+        layout="@layout/toolbar" />
 
-        <ImageView
-            android:id="@+id/image_empty"
-            android:layout_width="112dp"
-            android:layout_height="86dp"
-            android:layout_marginBottom="@dimen/margin_medium"
-            android:src="@drawable/penandink" />
-
-        <org.wordpress.android.widgets.WPTextView
-            android:id="@+id/title_empty"
-            style="@style/WordPress.EmptyList.Title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="@dimen/empty_list_title_bottom_margin"
-            android:layout_marginLeft="@dimen/empty_list_title_side_margin"
-            android:layout_marginRight="@dimen/empty_list_title_side_margin"
-            android:text="@string/empty_list_default"
-            app:fixWidowWords="true" />
-
-    </LinearLayout>
-
-    <android.support.design.widget.CoordinatorLayout
-        android:id="@+id/coordinator"
+    <RelativeLayout
+        android:id="@+id/root_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <org.wordpress.android.util.widgets.CustomSwipeRefreshLayout
-            android:id="@+id/ptr_layout"
+
+        <LinearLayout
+            android:id="@+id/empty_view"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:animateLayoutChanges="true"
+            android:background="@color/background_grey"
+            android:gravity="center"
+            android:orientation="vertical"
+            android:visibility="gone"
+            tools:visibility="visible">
+
+            <ImageView
+                android:id="@+id/image_empty"
+                android:layout_width="112dp"
+                android:layout_height="86dp"
+                android:layout_marginBottom="@dimen/margin_medium"
+                android:src="@drawable/penandink" />
+
+            <org.wordpress.android.widgets.WPTextView
+                android:id="@+id/title_empty"
+                style="@style/WordPress.EmptyList.Title"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginBottom="@dimen/empty_list_title_bottom_margin"
+                android:layout_marginLeft="@dimen/empty_list_title_side_margin"
+                android:layout_marginRight="@dimen/empty_list_title_side_margin"
+                android:text="@string/empty_list_default"
+                app:fixWidowWords="true" />
+
+        </LinearLayout>
+
+        <android.support.design.widget.CoordinatorLayout
+            android:id="@+id/coordinator"
             android:layout_width="match_parent"
             android:layout_height="match_parent">
 
-            <android.support.v7.widget.RecyclerView
-                android:id="@+id/recycler_view"
+            <org.wordpress.android.util.widgets.CustomSwipeRefreshLayout
+                android:id="@+id/ptr_layout"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:paddingTop="@dimen/margin_medium"
-                android:scrollbars="vertical" />
+                android:layout_height="match_parent">
 
-        </org.wordpress.android.util.widgets.CustomSwipeRefreshLayout>
+                <android.support.v7.widget.RecyclerView
+                    android:id="@+id/recycler_view"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:paddingTop="@dimen/margin_medium"
+                    android:scrollbars="vertical" />
 
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/fab_button"
+            </org.wordpress.android.util.widgets.CustomSwipeRefreshLayout>
+
+            <android.support.design.widget.FloatingActionButton
+                android:id="@+id/fab_button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="end|bottom"
+                android:layout_marginBottom="@dimen/fab_margin"
+                android:layout_marginRight="@dimen/fab_margin"
+                android:src="@drawable/gridicon_create_light"
+                app:borderWidth="0dp"
+                app:rippleColor="@color/fab_pressed" />
+
+        </android.support.design.widget.CoordinatorLayout>
+
+        <ProgressBar
+            android:id="@+id/progress"
+            style="?android:attr/progressBarStyleInverse"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="end|bottom"
-            android:layout_marginBottom="@dimen/fab_margin"
-            android:layout_marginRight="@dimen/fab_margin"
-            android:src="@drawable/gridicon_create_light"
-            app:borderWidth="0dp"
-            app:rippleColor="@color/fab_pressed" />
+            android:layout_alignParentBottom="true"
+            android:layout_centerHorizontal="true"
+            android:layout_marginBottom="@dimen/margin_large"
+            android:visibility="gone"
+            tools:visibility="visible" />
 
-    </android.support.design.widget.CoordinatorLayout>
+    </RelativeLayout>
 
-    <ProgressBar
-        android:id="@+id/progress"
-        style="?android:attr/progressBarStyleInverse"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="@dimen/margin_large"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-</RelativeLayout>
-
+</LinearLayout>
